+++ src/AppServices.Compliance/Enforcement/EnforcementActionCommand/AdministrativeOrderCommandDto.cs
@@ -17,7 +17,7 @@ public record AdministrativeOrderCommandDto : NotesDto
     [DisplayFormat(DataFormatString = DateTimeFormats.DateOnlyInput, ApplyFormatInEditMode = true)]
     [Display(Name = "Issued (Sent to Facility)")]
     [MaxDate]
-    public DateOnly? IssueDate { get; init; }
+    public DateOnly? IssueDate { get; set; }
 
     [DataType(DataType.Date)]
     [DisplayFormat(DataFormatString = DateTimeFormats.DateOnlyInput, ApplyFormatInEditMode = true)]