+++ src/AppServices.Compliance/Enforcement/EnforcementActionCommand/EnforcementActionReviewsDtos.cs
@@ -1,5 +1,7 @@
 ﻿using AirWeb.AppServices.Core.CommonDtos;
+using AirWeb.AppServices.Core.Utilities;
 using AirWeb.Domain.Compliance.EnforcementEntities.EnforcementActions.ActionProperties;
+using AirWeb.Domain.Core.Data.DataAttributes;
 using FluentValidation;
 
 namespace AirWeb.AppServices.Compliance.Enforcement.EnforcementActionCommand;
@@ -9,6 +11,13 @@ public record EnforcementActionRequestReviewDto
     [Required(ErrorMessage = "A reviewer must be selected.")]
     [Display(Name = "Request review from")]
     public string? RequestedOfId { get; init; }
+
+    [Required]
+    [DataType(DataType.Date)]
+    [DisplayFormat(DataFormatString = DateTimeFormats.DateOnlyInput, ApplyFormatInEditMode = true)]
+    [Display(Name = "Date Requested")]
+    [MaxDate]
+    public DateOnly RequestedDate { get; init; } = DateOnly.FromDateTime(DateTime.Today);
 }
 
 public class ReviewRequestValidator : AbstractValidator<EnforcementActionRequestReviewDto>
@@ -18,6 +27,10 @@ public class ReviewRequestValidator : AbstractValidator<EnforcementActionRequest
         RuleFor(dto => dto.RequestedOfId)
             .NotEmpty()
             .WithMessage("A reviewer must be selected.");
+
+        RuleFor(dto => dto.RequestedDate)
+            .Must(date => date <= DateOnly.FromDateTime(DateTime.Today))
+            .WithMessage("The Date Requested cannot be in the future.");
     }
 }
 
@@ -28,6 +41,12 @@ public record EnforcementActionSubmitReviewDto : NotesDto
 
     [Display(Name = "Request additional review from")]
     public string? RequestedOfId { get; init; }
+
+    [DataType(DataType.Date)]
+    [DisplayFormat(DataFormatString = DateTimeFormats.DateOnlyInput, ApplyFormatInEditMode = true)]
+    [Display(Name = "Date additional review requested")]
+    [MaxDate]
+    public DateOnly? RequestedDate { get; init; } = DateOnly.FromDateTime(DateTime.Today);
 }
 
 public class SubmitReviewValidator : AbstractValidator<EnforcementActionSubmitReviewDto>
@@ -41,5 +60,10 @@ public class SubmitReviewValidator : AbstractValidator<EnforcementActionSubmitRe
             .NotEmpty()
             .When(dto => dto.Result == ReviewResult.Forwarded)
             .WithMessage("A reviewer must be selected.");
+
+        RuleFor(dto => dto.RequestedDate)
+            .NotNull()
+            .When(dto => dto.Result == ReviewResult.Forwarded)
+            .WithMessage("Requested date must be set.");
     }
 }