+++ src/IaipDataService/IaipConstants.cs
@@ -2,6 +2,6 @@
 
 public static class IaipConstants
 {
-    // July 4, 1776, was originally used as a placeholder date for what should have been null values.
+    // July 4, 1776, was originally used in the IAIP as a placeholder date for what should have been null values.
     public static DateOnly NullPlaceholderDate { get; } = new(1776, 7, 4);
 }